+++ src/stratega/src/GUI/TileMap.cpp
@@ -21,7 +21,7 @@ namespace SGA
 
 		// resize the vertex array to fit the level size
 		m_vertices.setPrimitiveType(sf::Quads);
-		m_vertices.resize(state.getBoardWidth() * state.getBoardHeight() * 4);
+		m_vertices.resize(static_cast<size_t>(state.getBoardWidth() * state.getBoardHeight() * 4));
 	}
 
 	void TileMap::update(const GameState& state, const GameState& fowState, bool renderFow, FogRenderType fogType)
@@ -32,28 +32,28 @@ namespace SGA
 			for (int y = 0; y < state.getBoardHeight(); ++y)
 			{
 				// get the current tile
-				const auto& tile = state.getTileAt({x, y});
-				const auto& fogTile = fowState.getTileAt({x, y});
+				const auto& tile = state.getTileAt({ x, y });
+				const auto& fogTile = fowState.getTileAt({ x, y });
 
 				// get a pointer to the current tile's quad
-				sf::Vertex* quad = &m_vertices[(x + y * state.getBoardWidth()) * 4];
-				if(!renderFow)
+				sf::Vertex* quad = &m_vertices[static_cast<size_t>((x + y * state.getBoardWidth()) * 4)];
+				if (!renderFow)
 				{
 					updateTileQuad(quad, tile, sf::Color::White);
 				}
-				else if(fogType == FogRenderType::Nothing)
+				else if (fogType == FogRenderType::Nothing)
 				{
 					updateTileQuad(quad, tile, fogTile.getTileTypeID() == -1 ? sf::Color::Transparent : sf::Color::White);
 				}
-				else if(fogType == FogRenderType::Fog)
+				else if (fogType == FogRenderType::Fog)
 				{
 					updateTileQuad(quad, fogTile, sf::Color::White);
 				}
-				else if(fogType == FogRenderType::Tiles)
+				else if (fogType == FogRenderType::Tiles)
 				{
 					updateTileQuad(quad, tile, fogTile.getTileTypeID() == -1 ? FOG_COLOR : sf::Color::White);
 				}
-				
+
 			}
 		}
 	}
@@ -61,13 +61,13 @@ namespace SGA
 	void TileMap::updateTileQuad(sf::Vertex* quadPtr, const Tile& tile, const sf::Color& quadColor) const
 	{
 		// define the 4 corners
-		auto start = toISO(tile.x(), tile.y()) - sf::Vector2f{TILE_ORIGIN_X, TILE_ORIGIN_Y};
+		auto start = toISO(tile.x(), tile.y()) - sf::Vector2f{ TILE_ORIGIN_X, TILE_ORIGIN_Y };
 		auto tileSize = tileset.getSpriteSize();
 		quadPtr[0].position = sf::Vector2f(start.x, start.y);
-		quadPtr[1].position = sf::Vector2f(start.x + tileSize.x, start.y);
-		quadPtr[2].position = sf::Vector2f(start.x + tileSize.x, start.y + tileSize.y);
-		quadPtr[3].position = sf::Vector2f(start.x, start.y + tileSize.y);
-		
+		quadPtr[1].position = sf::Vector2f(start.x + static_cast<float>(tileSize.x), start.y);
+		quadPtr[2].position = sf::Vector2f(start.x + static_cast<float>(tileSize.x), start.y + static_cast<float>(tileSize.y));
+		quadPtr[3].position = sf::Vector2f(start.x, start.y + static_cast<float>(tileSize.y));
+
 		// define the 4 texture coordinates
 		auto rect = tileset.getSpriteRect(tileSpritePaths.at(tile.getTileTypeID()));
 		quadPtr[0].texCoords = sf::Vector2f(rect.left, rect.top);
